將 sort list 邏輯作為 model 獨立出
